pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = ["Tran Viet Phuoc <phuoc.finn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.7.0"
Flask-Script = "^2.0.6"
Flask-Login = "^0.5.0"
Flask-Mail = "^0.9.1"
Flask-WTF = "^0.14.3"
WTForms = "^2.3.3"
black = "^20.8b1"
pynvim = "^0.4.3"
Flask-Admin = "^1.5.7"
elasticsearch = "^7.11.0"
Pillow = "^8.1.2"
jedi = "^0.18.0"
flake8 = "^3.8.4"
python-dotenv = "^0.15.0"
email-validator = "^1.1.2"
psycopg2 = "^2.8.6"
pytest = "^6.2.4"
Flask-Security = "^3.0.0"
Flask = "1.1.4"
Flask-Babel = "^2.0.0"
marshmallow-sqlalchemy = "^0.26.1"
apispec = "^5.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
